requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
PyYAML==6.0.2
python-dotenv==1.0.1
prometheus-fastapi-instrumentator==6.1.0
